Resolve Compiler Warnings (#64)

+++ Source/Core/Editor/Utils/ERS_Editor_ThemeManager/ERS_Editor_ThemeManager.cpp
@@ -15,7 +15,7 @@ ERS_CLASS_ThemeManager::ERS_CLASS_ThemeManager(ERS_CLASS_LoggingSystem* Logger, 
 
     // Default To Dark Mode
     int Index;
-    for (Index = 0; Index < ThemeNames_.size(); Index++) {
+    for (Index = 0; (long)Index < ThemeNames_.size(); Index++) {
         if (ThemeNames_[Index] == "Dark") {
             break;
         }
